email-api/requirements.txt
fastapi==0.115.0
mangum==0.19.0
resend==0.8.0
python-dotenv==1.0.1
